# [move-lang] rewrote alias scoping, improving compiler perf by ~65%