--- a/language/move-lang/tests/move_check/expansion/use_function_tparam_shadows.exp
+++ b/language/move-lang/tests/move_check/expansion/use_function_tparam_shadows.exp
@@ -1,3 +1,9 @@
+warning[W09001]: unused alias
+  ┌─ tests/move_check/expansion/use_function_tparam_shadows.move:7:17
+  │
+7 │     use 0x2::X::foo;
+  │                 ^^^ Unused 'use' of alias 'foo'. Consider removing it
+
 error[E03005]: unbound unscoped name
    ┌─ tests/move_check/expansion/use_function_tparam_shadows.move:10:9
    │