- Old alias scoping was done with an iter+filter to do set difference
- New alias scoping instead keeps track of unused items at each scope
- For each new scope, the new algorithm fully populates a set of "unused" aliases. Each usage of an alias then removes that from the set, so at the end of the scope, it can instantaneously answer what aliases are unused (as opposed to doing set difference)
- Additionally, clones are reduced when the scope does not introduce anything new. This accounted for about an additional 11% improvement on release builds

## Motivation

- According to flamegraph, expansion was taking 60% to 70% of the total runtime for the compiler
- This motivated the rewrite and led to some nice perf gains!
```
dev build DF+Std (averaged over 10 iterations)
OLD       8.155s
NEW       2.840s
REDUCTION 65.17%

release build DF+Std (averaged over 100 iterations)
OLD       0.849s
NEW       0.271s
REDUCTION 68.08%

nextest -p move-lang
OLD       17.623s
NEW       16.102s
REDUCTION 8.63%

nextest -p move-lang-functional-tests
OLD       511.871s
NEW       185.475s
REDUCTION 63.77%

nextest -p move-prover
OLD       650.349s
NEW       287.221s
REDUCTION 55.84%

x test -p move-lang
OLD       10.047s
NEW       7.335s
REDUCTION 26.99%

x test -p move-lang-functional-tests
OLD       51.223s
NEW       42.450s
REDUCTION 17.13%

x test -p move-prover
OLD       534.380s
NEW       228.259s
REDUCTION 57.29%

```

## Test Plan

- Actually found a bug while doing this too! Updated that test